+++ internal/platform/platform_unix.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package platform // import "github.com/docker/docker/pkg/platform"
+package platform
 
 import (
 	"golang.org/x/sys/unix"
@@ -13,4 +13,13 @@ func runtimeArchitecture() (string, error) {
 		return "", err
 	}
 	return unix.ByteSliceToString(utsname.Machine[:]), nil
+}
+
+// NumProcs returns the number of processors on the system
+//
+// Deprecated: temporary stub for non-Windows to provide an alias for the deprecated github.com/docker/docker/pkg/platform package.
+//
+// FIXME(thaJeztah): remove once we remove  github.com/docker/docker/pkg/platform
+func NumProcs() uint32 {
+	return 0
 }
